feat(#622): catalog prices declare explicit access windows (hard cut)
Replace the billing_cycle_days + intro price model with an access-window
model. A price now declares:
  duration   - access window a purchase grants ("30d", "indefinite", ...);
               omitted = indefinite/durable
  auto_renew - whether it recharges + extends after duration (a subscription);
               requires a finite duration
  trial      - optional first phase {unit_amount, duration} (was intro)

This lets one product carry both recurring and one-off finite-window prices
(e.g. doujins legacy Solana premium: $23/30d and $62/90d one-offs alongside the
recurring tier) without forking into duplicate products — auto_renew is part of
price identity, so $23/30d recurring and $23/30d one-off no longer collide.

Hard cut, no aliases:
- migration 037 drops billing_cycle_days/initial_amount/initial_period_days,
  adds access_duration_days/auto_renew/trial_unit_amount/trial_duration_days
  with checks (auto_renew needs a duration; trial needs auto_renew) and an
  auto_renew-aware unique constraint; sqlc regenerated.
- models.Price: AccessDurationDays/AutoRenew/Trial*, RecurringCycleDays()/
  IsRecurring() helpers replace BillingCycleDays at recurring-cadence callers.
- catalog manifest grammar: duration/auto_renew/trial replace interval/intro.
- checkout: one-off finite windows now set BOTH entitlement end_at and
  product_access ends_at from the price's access window (G2/G3); product-level
  per-entitlement durations still apply when the price is indefinite. Coverage
  stacking unchanged.
- Solana adapter no-ops (no on-chain plan) for one-off prices.
- example catalog + full test suite updated to the new grammar.

+++ internal/db/queries/payments.sql
@@ -242,7 +242,7 @@ WHERE p.subscription_id IS NOT NULL
   AND p.rail = sqlc.arg(rail)
   AND sub.rail::text = p.rail::text
   AND p.amount > 0
-  AND p.amount = sqlc.arg(amount_cents) * 10000
+  AND p.amount = sqlc.arg(amount_cents)::bigint * 10000
   AND RIGHT(regexp_replace(COALESCE(pm.last_four, ''), '[^0-9]', '', 'g'), 4) = sqlc.arg(last4)::text
   AND p.purchased_at >= sqlc.arg(from_at)::timestamptz
   AND p.purchased_at <= sqlc.arg(to_at)::timestamptz
